Detach configs to avoid Circular Reference
