+++ src/main/java/com/chordplay/chordplayapiserver/infra/redis/RedisListenerConfig.java
@@ -1,0 +1,30 @@
+package com.chordplay.chordplayapiserver.infra.redis;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnection;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.listener.ChannelTopic;
+import org.springframework.data.redis.listener.RedisMessageListenerContainer;
+import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
+
+@Configuration
+@RequiredArgsConstructor
+public class RedisListenerConfig {
+
+    private final RedisConnectionFactory redisConnectionFactory;
+    private final RedisSubscriber redisSubscriber;
+    @Bean
+    RedisMessageListenerContainer redisContainer() {
+        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
+        container.setConnectionFactory(redisConnectionFactory);
+        container.addMessageListener(messageListenerAdapter(), new ChannelTopic("ch01"));
+        return container;
+    }
+
+    @Bean
+    MessageListenerAdapter messageListenerAdapter() {
+        return new MessageListenerAdapter(redisSubscriber);
+    }
+}
